BE: [refactor] PostService 기능 분리 #36
