+++ src/backend/src/main/java/RunningMachines/R2R/domain/community/post/controller/PostController.java
@@ -3,6 +3,7 @@ package RunningMachines.R2R.domain.community.post.controller;
 import RunningMachines.R2R.domain.community.post.dto.PostCreateRequestDto;
 import RunningMachines.R2R.domain.community.post.dto.PostShowDetailResponseDto;
 import RunningMachines.R2R.domain.community.post.service.PostCommandService;
+import RunningMachines.R2R.domain.community.post.service.PostQueryService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -12,6 +13,7 @@ import org.springframework.web.bind.annotation.*;
 @RequiredArgsConstructor
 public class PostController {
     private final PostCommandService postCommandService;
+    private final PostQueryService postQueryService;
 
     @PostMapping("/{boardName}")
     public ResponseEntity<Long> createPost(@PathVariable String boardName, @RequestBody PostCreateRequestDto postCreateRequestDto) {
@@ -21,7 +23,7 @@ public class PostController {
 
     @GetMapping("/{boardName}/{postId}")
     public ResponseEntity<PostShowDetailResponseDto> getPostWithComments(@PathVariable String boardName, @PathVariable Long postId) {
-        PostShowDetailResponseDto responseDto = postCommandService.getPostWithComments(postId);
+        PostShowDetailResponseDto responseDto = postQueryService.getPostWithComments(postId);
         return ResponseEntity.ok(responseDto);
     }
 }
