+++ src/backend/src/main/java/RunningMachines/R2R/domain/community/post/service/PostQueryService.java
@@ -1,14 +1,23 @@
 package RunningMachines.R2R.domain.community.post.service;
 
-
+import RunningMachines.R2R.domain.community.post.dto.PostShowDetailResponseDto;
+import RunningMachines.R2R.domain.community.post.entity.Post;
 import RunningMachines.R2R.domain.community.post.repository.PostRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @Slf4j
 @Service
 @RequiredArgsConstructor
 public class PostQueryService {
     private final PostRepository postRepository;
+
+    @Transactional(readOnly = true)
+    public PostShowDetailResponseDto getPostWithComments(Long postId) {
+        Post post = postRepository.findById(postId)
+                .orElseThrow(() -> new IllegalArgumentException("게시글을 찾을 수 없습니다."));
+        return PostShowDetailResponseDto.fromPost(post);
+    }
 }
